Added zom -> Zombies fix for T4ZM
Both cmp and zom are now handled as the "Zombies" gametype.

It was observed that after the recent Pluto T4 refactor, some servers began reporting the gametype as both cmp and zom. 

While im unsure if zom is related to the refactor or not, but this change ensures both are correctly identified.